[mesh] Add new mesh crate
I need to work on the mesh data structure before trying to add
any slicer logic. For now we just have a data structure that
stores facets as indices into a list of vectors. This is simple
but I am expecting I can put most of the interesting behavior
behind a trait so that we can decouple the specific mesh data
structure from the code that is consuming it.