[slicer] Drop superfluous points from contours
We can get extra points in our contour caused by the mesh
representation of the model. Since STL files can only represent
triangles, this is obvious with even the simple case of a square
face, which could be represented as:

  *-----*
  |\    |
  | \   |
  |  \  |
  |   \ |
  |    \|
  *-----*

If this geometry would intersect the cutting plane, we would get
an additional intersection point from the diagonal:

  *-----*
  |\    |
  | \   |
--o--o--o-- Cutting Plane
  |   \ |
  |    \|
  *-----*

When building the contour we can drop points if we determine they
are not needed.