[slicer][svg] Add svg grid output
This works for some basic cases, but at least on chrome only content
'above the fold' is drawn, which means we get the first 60-70 frames
of a 20mm calibration cube with default parameters.

With this it's easy to see we have an issue where the first layer
path is reversed from the rest of the layers. Most likely because
we are adding those path segments with the triangles on the
horizontal faces whereas later layers are using triangles from the
vertical faces.