[slicer] Always compute segment direction
We had logic to determine the direction of intersection when two edges
of a triangle intersect with the cutting plane, but we were not using
that logic when a triangle edge lies on the cutting plane.

This fixes an inssue where the first layer path runs backwards from
other paths.